refactor: streamline import statements and enhance TransferBase functionality with recurrence details
